Move pixel rounding into its own file
